ENH:Si al falla la desarrega d'un estudi l'esborrem del disc dur i de la base de dades per si hi estava inserit, també es millora el mètode per esborrar l'estudi només en els errors en els quals es pot haver descarregat algun fitxer

+++ starviewer/src/inputoutput/qexecuteoperationthread.cpp
@@ -408,7 +408,24 @@ void QExecuteOperationThread::errorRetrieving(QString studyInstanceUID, QString 
 
     emit errorInOperation(studyInstanceUID, pacsID, lastError);
 
-    deleteDirectory.deleteDirectory(localDatabaseManager.getStudyPath(studyInstanceUID), true);
+	INFO_LOG("Esborrem els fitxers del disc i de la base de dades l'estudi " + studyInstanceUID + " que s'ha descarregat incorrectament");
+	//TODO:Esborrar només els fitxers que s'hagin pogut descarregar, no tot l'estudi
+
+	if (lastError == QExecuteOperationThread::MoveFailureOrRefusedStatus || lastError == QExecuteOperationThread::MoveUnknowStatus ||
+		lastError == QExecuteOperationThread::DatabaseError || lastError == QExecuteOperationThread::PatientInconsistent)
+	{
+		//Si s'ha produït un error dels tipus anteriors potser que tinguem fitxers descarregats, per tant esborrem l'estudi
+
+		deleteDirectory.deleteDirectory(localDatabaseManager.getStudyPath(studyInstanceUID), true);
+
+		/* Imaginem el cas que descarreguem la sèrie 1 del estudi correctament i és inserit a la base de dades, llavors intentem descarregar la sèrie 2
+		 * però aquesta falla, no es pot descarregar, en aquest s'executa aquest mètode i es fa una crida a deleteDirectory, esborrant qualsevol fitxer
+		 * que s'hagi pogut descarregar d'aquesta descarrega fallida i els de la serèi 1 que si s'havien descarregat correctament, perquè el deleteDirectory esborra
+		 * tots els fitxers de l'estudi, al passar això si no esborréssim l'estudi de la base de dades quedaria inconsistent perquè tindríem l'estudi amb la
+		 * sèrie 1 a la base de dades, però els fitxers estarien esborrats.
+		 */
+		localDatabaseManager.del(studyInstanceUID);
+	}
 }
 
 void QExecuteOperationThread::seriesRetrieved(QString studyInstanceUID)
